LPS-72928 Extract spring extender component scanning to new module
