+++ modules/apps/foundation/portal-osgi-debug/portal-osgi-debug-spring-extender/src/main/java/com/liferay/portal/osgi/debug/spring/extender/internal/SpringExtenderConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.spring.extender.internal.configuration;
+package com.liferay.portal.osgi.debug.spring.extender.internal;
 
 import aQute.bnd.annotation.metatype.Meta;
