--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_ar.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=التكوين الموسع الربيع (Automatic Translation)
-unavailable-component-scanning-interval-help=تعيين المكون غير متوفر المسح الفاصل الزمني بالثانية. عند تعيين إلى صفر أو سلبي، سيتم تعطيل المسح الضوئي. القيمة الافتراضية هي-1. (Automatic Translation)
