--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_bg.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Пролет Extender конфигурация (Automatic Translation)
-unavailable-component-scanning-interval-help=Задайте недостъпни компонент сканиране интервал от секунда. Когато зададете нула или отрицателни, сканирането ще бъде забранена. Стойността по подразбиране е -1. (Automatic Translation)
