--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_ca.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Configuració Exoferta de primavera (Automatic Translation)
-unavailable-component-scanning-interval-help=Establir el component disponible interval d'escaneig per segon. Quan posa a zero o negatives, l'exploració es desactivarà. El valor per defecte és -1. (Automatic Translation)
