--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_cs.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Konfigurace zařízení Extender jaro (Automatic Translation)
-unavailable-component-scanning-interval-help=Nastavení k dispozici komponentu interval odečítání o sekundu. Pokud nastavíte hodnotu nula nebo záporné číslo, skenování bude zakázán. Výchozí hodnota je -1. (Automatic Translation)
