--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_el.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Ρύθμιση παραμέτρων επέκτασης άνοιξη (Automatic Translation)
-unavailable-component-scanning-interval-help=Ορισμός στοιχείου διαθέσιμος διάστημα σάρωσης από το δεύτερο. Όταν οριστεί σε μηδέν ή αρνητικό, θα είναι απενεργοποιημένη η σάρωση. Η προεπιλεγμένη τιμή είναι -1. (Automatic Translation)
