--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_es.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Configuración del extensor de primavera (Automatic Translation)
-unavailable-component-scanning-interval-help=Establecer el componente de carácter análisis intervalo por segundo. Cuando se establece a cero o negativo, se deshabilitará la exploración. El valor predeterminado es -1. (Automatic Translation)
