--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_et.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Kevadel ekstender konfiguratsiooni (Automatic Translation)
-unavailable-component-scanning-interval-help=Seadke intervalli skaneerimine läbi teise kättesaamatuks komponent. Kui valitud null või negatiivne, keelatakse skannimine. Vaikimisi väärtus on -1. (Automatic Translation)
