--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_fa.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=بهار توسعه دهنده پیکربندی (Automatic Translation)
-unavailable-component-scanning-interval-help=مجموعه جزء موجود اسکن فاصله توسط دوم. هنگام تنظیم روی صفر یا منفی اسکن غیرفعال خواهد شد. مقدار پیش فرض -1 است. (Automatic Translation)
