--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_fr.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Configuration de Spring Extender (Automatic Translation)
-unavailable-component-scanning-interval-help=Définir le composant indisponible intervalle de balayage par seconde. Lorsque la valeur zéro ou négatif, le scan sera désactivé. La valeur par défaut est -1. (Automatic Translation)
