--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_hi_IN.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=वसंत भरनेवाला कॉन्फ़िगरेशन (Automatic Translation)
-unavailable-component-scanning-interval-help=अनुपलब्ध घटक द्वारा दूसरे अंतराल स्कैनिंग को सेट करें। जब शून्य या नकारात्मक करने के लिए सेट, स्कैन अक्षम कर दिया जाएगा। डिफ़ॉल्ट मान -1 है। (Automatic Translation)
