--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_hu.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Tavaszi Extender konfiguráció (Automatic Translation)
-unavailable-component-scanning-interval-help=Mérési időközök második elérhetetlen összetevő beállítása Ha értéke nulla vagy negatív, a vizsgálatot le lesz tiltva. Az alapértelmezett érték -1. (Automatic Translation)
