--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_in.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Musim semi Extender konfigurasi (Automatic Translation)
-unavailable-component-scanning-interval-help=Membuat tersedia komponen pemindaian interval oleh kedua. Bila diatur ke nol atau negatif, scan akan dinonaktifkan. Nilai default adalah-1. (Automatic Translation)
