--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_it.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Configurazione del dispositivo Extender di primavera (Automatic Translation)
-unavailable-component-scanning-interval-help=Impostare il componente non disponibile intervallo di scansione per secondo. Quando impostato a zero o negativo, la scansione verrà disabilitata. Il valore predefinito è -1. (Automatic Translation)
