--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_iw.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=תצורת ה-Extender האביב (Automatic Translation)
-unavailable-component-scanning-interval-help=להגדיר את הרכיב לא זמינה לסריקות מרווח השנייה. כאשר מוגדר אפס או מספר שלילי, הסריקה לא יהיה זמין. ערך ברירת המחדל הוא-1. (Automatic Translation)
