--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_ko.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=봄 Extender 구성 (Automatic Translation)
-unavailable-component-scanning-interval-help=초 간격을 검색 하는 사용할 수 없는 구성 요소를 설정 합니다. 0 또는 음수를 설정 하면, 검색 비활성화 됩니다. 기본 값은-1입니다. (Automatic Translation)
