--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_lt.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Pavasarį plėstuvo konfigūracijos (Automatic Translation)
-unavailable-component-scanning-interval-help=Nustatyti nuskaitymo intervalas pagal antrą nepasiekiamas komponentas. Nustačius į neigiamas arba lygus nuliui, nuskaitymas bus išjungtas. Numatytoji reikšmė yra -1. (Automatic Translation)
