--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_nb.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Våren Extender-konfigurasjon (Automatic Translation)
-unavailable-component-scanning-interval-help=Angi komponenten utilgjengelig skanning intervall av andre. Når satt til null eller negativ, vil skanningen bli deaktivert. Standardverdien er -1. (Automatic Translation)
