--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_nl.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Spring Extender Configuration (Automatic Copy)
-unavailable-component-scanning-interval-help=Set the unavailable component scanning interval by second. When set to zero or negative, the scan will be disabled. The default value is -1. (Automatic Copy)
