--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_pl.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Konfiguracja urządzenia Extender wiosna (Automatic Translation)
-unavailable-component-scanning-interval-help=Ustaw ten składnik niedostępny przez sekundę między pomiarami. Gdy zestaw do zera lub ujemne, skanowania zostanie wyłączona. Wartością domyślną jest -1. (Automatic Translation)
