--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_pt_PT.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Configuração de extensor de primavera (Automatic Translation)
-unavailable-component-scanning-interval-help=Defina o componente indisponível digitalização intervalo por segundo. Quando definido como zero ou negativo, o exame será desativado. O valor padrão é -1. (Automatic Translation)
