--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_ro.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Configurare de prelungire de primăvară (Automatic Translation)
-unavailable-component-scanning-interval-help=Seta componenta indisponibil intervalul de scanare de al doilea. Când este setat la zero sau negativ, scanare vor fi dezactivate. Valoarea implicită este -1. (Automatic Translation)
