--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_ru.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Весной расширитель конфигурации (Automatic Translation)
-unavailable-component-scanning-interval-help=Установите компонент недоступен, сканирования интервала на второй. Если задано нулевое или отрицательное, проверка будет отключена. Значение по умолчанию равно -1. (Automatic Translation)
