--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_sk.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Konfigurácia zariadenia Extender jar (Automatic Translation)
-unavailable-component-scanning-interval-help=Nastavenie nedostupné súčasti interval snímania druhou. Keď nastavené na nulu alebo záporné, scan bude vypnuté. Predvolená hodnota je -1. (Automatic Translation)
