--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_sl.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Spomladi Extender konfiguracijo (Automatic Translation)
-unavailable-component-scanning-interval-help=Nastavite na voljo komponente skeniranje interval drugi. Ko je nastavljena na nič ali negativno vrednost, skandirati hoteti obstati nesposoben. Privzeta vrednost je -1. (Automatic Translation)
