--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_th.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=การกำหนดค่า Extender สปริง (Automatic Translation)
-unavailable-component-scanning-interval-help=ตั้งค่าคอมโพเนนต์ที่พร้อมใช้งานการสแกนช่วง โดยสอง เมื่อตั้งค่าเป็นศูนย์ หรือติดลบ การสแกนจะทำงาน ค่าเริ่มต้นคือ -1 (Automatic Translation)
