--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_tr.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Bahar Extender'ı yapılandırma (Automatic Translation)
-unavailable-component-scanning-interval-help=Her saniye aralığı tarama kullanılamıyor bileşen ayarlayın. Sıfır veya negatif olarak ayarlandığında, tarama devre dışı bırakılacak. Varsayılan değer -1'dir. (Automatic Translation)
