--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_uk.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Весна Extender конфігурації (Automatic Translation)
-unavailable-component-scanning-interval-help=Встановити недоступний компонент сканування інтервал за секунду. Якщо встановлено нулю або від'ємні, сканування буде вимкнуто. За промовчанням значення-1. (Automatic Translation)
