--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_vi.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=Mùa xuân mở rộng cấu hình (Automatic Translation)
-unavailable-component-scanning-interval-help=Đặt các thành phần không có sẵn chức năng quét các khoảng thời gian bằng cách thứ hai. Khi thiết lập để zero hoặc phủ định, quá trình quét sẽ bị vô hiệu. Giá trị mặc định là -1. (Automatic Translation)
