--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_zh_CN.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=春天扩展程序配置 (Automatic Translation)
-unavailable-component-scanning-interval-help=设置不可用的组件，由第二次扫描间隔。当设置为零或负数，扫描将被禁用。默认值为-1。 (Automatic Translation)
