--- modules/apps/foundation/portal/portal-spring-extender/src/main/resources/content/Language_zh_TW.properties
@@ -1,2 +1,0 @@
-spring-extender-configuration-name=春天擴展程式配置 (Automatic Translation)
-unavailable-component-scanning-interval-help=設置不可用的元件，由第二次掃描間隔。當設置為零或負數，掃描將被禁用。預設值為-1。 (Automatic Translation)
